requirements.txt
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.0.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.3.0
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.1.0
pytz==2023.3
reportlab==4.0.4
requests==2.32.5
SQLAlchemy==1.4.53
urllib3==2.6.3
Werkzeug==2.3.7
WTForms==3.0.1
google-generativeai==0.8.3
groq==0.11.0
python-dotenv
Flask-Limiter
psycopg2-binary
